+++ ch04/rtda/operand_stack.go
@@ -32,8 +32,8 @@ func (self *OperandStack) PushLong(val int64) {
 }
 func (self *OperandStack) PopLong() int64 {
 	self.size -= 2
-	high := self.slots[self.size+1].num
-	lower := self.slots[self.size].num
+	high := uint32(self.slots[self.size+1].num)
+	lower := uint32(self.slots[self.size].num)
 	return int64(high)<<32 | int64(lower)
 }
